feat: added the ability to edit profiles with specific percentages

+++ src/serialization/profile/ProfileData.hpp
@@ -5,6 +5,6 @@
 
 struct ProfileData
 {
-  std::vector<int> tags;
+  std::vector<float> tags;
   std::vector<Stage> stages;
 };
